% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_functions.R
\name{build_estimated_fn}
\alias{build_estimated_fn}
\title{Build an estimated function that captures the estimated parameters}
\usage{
build_estimated_fn(prefix, dist_name, fn_estimate_params, x_data, noverlap = 1)
}
\arguments{
\item{prefix}{The prefix (generally one of \code{p}, \code{q}, \code{r} or \code{d}).}

\item{dist_name}{The name of the distribution.}

\item{fn_estimate_params}{A function that takes the data and the extent of the overlap in the data,
and returns a single object holding estimated parameters of the distribution being fitted.
The method of estimation should be unbiased. Note that for many distributions, MLE
only gives \emph{asymptotically} unbiased parameters. Users should validate that their estimation
functions are unbiased and if necessary adjust the threshold p-value to compensate for this.}

\item{x_data}{The data from which the parameters will be estimated.}

\item{noverlap}{The extent of any overlap in the data. \code{1} means no overlap.}
}
\value{
A function with captured estimated parameters that is called with a single argument
(e.g. \code{q} for a \code{p} function).
}
\description{
Build an estimated function that captures the estimated parameters
}
\examples{
x_data <- rnorm(100)
pnorm_est <- build_estimated_fn("p", "norm", estimate_mean_sd_ol, x_data)
pnorm_est(0)
pnorm(0, mean(x_data), sd(x_data))
}
