% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_functions.R
\name{build_uniparam_args}
\alias{build_uniparam_args}
\title{Build a uniparameter-friendly list of arguments}
\usage{
build_uniparam_args(x, params)
}
\arguments{
\item{x}{The non-parameter element of the arguments (e.g. \code{p}, for \code{q}-functions).}

\item{params}{The parameters.}
}
\value{
A list consisting of \code{x} and \code{params} in a format suitable for passing to \code{\link{do.call}}.
}
\description{
In general params will be a list returned by an estimation function, so just tack the x argument
on the front for do.call purposes later. However sometimes the params may be a bespoke object
of some kind, e.g. the result of calling ghyp() from the ghyp package, and needs to be wrapped in a list.
}
\examples{
build_uniparam_args(1:10, list(mean = 0, sd = 1))
}
