% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_functions.R
\name{build_uniparam_fn}
\alias{build_uniparam_fn}
\title{Build a uniparameter-friendly function}
\usage{
build_uniparam_fn(prefix, dist_name, fn = NULL)
}
\arguments{
\item{prefix}{The prefix (generally one of \code{p}, \code{q}, \code{r} or \code{d}).}

\item{dist_name}{The name of the distribution.}

\item{fn}{If not \code{NULL}, will override the result of calling \link{get_prefixed_dist_fn}
with \code{prefix} and \code{dist}}
}
\value{
A uniparameter version of the input function.
}
\description{
If \code{fn} is \code{NULL}, \link{get_prefixed_dist_fn} will be called to get an function.
This is then converted to a uniparameter-friendly version that calls \link{build_uniparam_args}
internally so it can then call \link{do.call}.
}
\examples{
uniparam_pnorm <- build_uniparam_fn("p", "norm")
uniparam_pnorm(0, list(mean = 0, sd = 1))
}
