% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{calc_ad_test_stat}
\alias{calc_ad_test_stat}
\title{Calculate the AD test statistic of data against a fitted distribution}
\usage{
calc_ad_test_stat(x, params, fn_p)
}
\arguments{
\item{x}{The data.}

\item{params}{The parameters of the distribution, generally estimated from the data,
as a single object.}

\item{fn_p}{The cumulative distribution function of the distribution,
in the form of a function that takes the data and estimated parameters as a single object,
and returns cumulative probability values.}
}
\value{
The AD test statistic.
}
\description{
The purpose of this function is to allow the AD test statistic to be calculated
for a wide range of distributions, i.e. to abstract the calculation of the test
statistic from the CDF (p) function.
This requires the estimated parameters to be specified as a single object.
}
\examples{
set.seed(1)
x <- rnorm(100)
params <- list(mean = mean(x), sd = sd(x))
calc_ad_test_stat(x, params, function(x, params) pnorm(x, params$mean, params$sd))
if (require(ADGofTest)) ADGofTest::ad.test(x, pnorm, mean(x), sd(x))$statistic
}
\seealso{
Other Test statistic functions: 
\code{\link{calc_ks_test_stat}()}
}
\concept{Test statistic functions}
