% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping_data.R
\name{calc_theo_sercor}
\alias{calc_theo_sercor}
\title{Calculate theoretical serial correlation at a given lag from overlapped data}
\usage{
calc_theo_sercor(noverlap, lag)
}
\arguments{
\item{noverlap}{The extent of the overlap, e.g. 12 means annual overlaps from monthly data.
1 means no overlap.}

\item{lag}{The lag in time between two overlapped observations.
The serial correlation at lags equal to or higher than \code{noverlap} is zero.}
}
\value{
The theoretical serial correlation.
}
\description{
Serial correlation arises from overlapped data because some of the information is 'shared'
between successive overlapping observations. The amount of serial correlation
is a function of the lag between observations and the extent of the overlap.
For example the serial correlation between Jan-Dec and Feb-Jan observations is higher
than between Jan-Dec and Sep-Aug-next-year observations.
}
\details{
The formula follows from writing each overlapping observation as the sum of
\code{noverlap} independent observations and counting the extent to which these two sets overlap.
Where the lag is higher than the overlap no information is shared and the theoretical
serial correlation becomes zero.
}
\examples{
calc_theo_sercor(12, 0:12)
}
\seealso{
Other Overlapping data functions: 
\code{\link{build_theo_sercor_mtx}()}
}
\concept{Overlapping data functions}
