% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{estimate_exp_ol}
\alias{estimate_exp_ol}
\title{Estimate the exponential distribution rate parameter by method of moments}
\usage{
estimate_exp_ol(x, noverlap = 1)
}
\arguments{
\item{x}{The data.}

\item{noverlap}{The extent to which the data is overlapped. This is not used,
but required for compatibility with \code{\link{gof_test_sim}}.}
}
\value{
A list with one item:
\describe{
\item{rate}{the rate parameter, calculated as the reciprocal of the mean of \code{x}}
}
}
\description{
This function has the right signature to be used in \code{\link{gof_test_sim}},
and estimates the rate parameter as the reciprocal of the sample mean.
}
\examples{
estimate_exp_ol(rexp(100))
}
\seealso{
Other Estimation functions: 
\code{\link{estimate_logis_ol}()},
\code{\link{estimate_mean_sd_ol}()}
}
\concept{Estimation functions}
