% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{estimate_mean_sd_ol}
\alias{estimate_mean_sd_ol}
\title{Estimate mean and standard deviation by method of moments allowing for potentially overlapping data}
\usage{
estimate_mean_sd_ol(x, noverlap = 1)
}
\arguments{
\item{x}{The data.}

\item{noverlap}{The extent to which the data is overlapped. \code{noverlap == 1} (the default) means no overlap,
and in this case \code{\link{calc_sd_ol_bias_fac}} returns 1, so no adjustment is applied.}
}
\value{
A list with two items:
\describe{
\item{mean}{the mean of \code{x}}
\item{sd}{the standard deviation of \code{x} adjusted for overlapping data bias, if any}
}
}
\description{
This function serves two purposes: (1) collate the mean and standard deviation into a single list object,
and (2) allow for bias in the standard deviation arising from the overlapping data
by using \code{\link{calc_sd_ol_bias_fac}}.
}
\examples{
estimate_mean_sd_ol(rnorm(100))
}
\seealso{
Other Estimation functions: 
\code{\link{estimate_exp_ol}()},
\code{\link{estimate_logis_ol}()}
}
\concept{Estimation functions}
