% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping_data.R
\name{gen_ol_rnorm_mtx}
\alias{gen_ol_rnorm_mtx}
\title{Generate a matrix of overlapped Normal variates}
\usage{
gen_ol_rnorm_mtx(nsims, nsteps, noverlap, mu = 0, sigma = 1)
}
\arguments{
\item{nsims}{The number of simulations (rows in the matrix).}

\item{nsteps}{The number of steps (columns in the matrix).}

\item{noverlap}{The extent of the overlap, e.g. 12 means annual overlaps from monthly data.
1 means no overlap.}

\item{mu}{The mean of the Normal distribution.}

\item{sigma}{The standard deviation of the Normal distribution.}
}
\value{
A \code{nsims} by \code{steps} matrix with overlapped values.
}
\description{
The function operates by calculating a matrix of 'pthly' variates
having 1/pth the mean and standard deviation, and then scanning each row and
summing the row values in overlapping sets of \code{p}.
}
\examples{
gen_ol_rnorm_mtx(100, 10, noverlap = 12)
}
