% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder_functions.R
\name{get_prefixed_dist_fn}
\alias{get_prefixed_dist_fn}
\title{Get a prefixed distribution function}
\usage{
get_prefixed_dist_fn(prefix, dist_name)
}
\arguments{
\item{prefix}{The prefix (generally one of \code{p}, \code{q}, \code{r} or \code{d}).}

\item{dist_name}{The name of the distribution.}
}
\value{
The function corresponding to the prefix and distribution_name.
}
\description{
Helper function to get, say, the function \code{\link[stats]{pnorm}}
by passing \code{"p"} and \code{"norm"}.
}
\examples{
get_prefixed_dist_fn("p", "norm")
}
