% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practechniques.R
\docType{package}
\name{practechniques}
\alias{practechniques}
\title{practechniques - Practical Actuarial Techniques}
\description{
A companion R package for the online book Practical Actuarial Techniques (in development).
}
\details{
THIS IS A PRE-RELEASE VERSION AND NO RELIANCE WHATSOEVER MAY BE PLACED ON IT.

At the time of writing (May 2022) the main feature is a framework
using Monte Carlo simulation to calculate p-values for statistical goodness-of-fit tests
where the parameters have been estimated from the data.

This is a generalisation of the approach in the
\href{https://cran.r-project.org/web/packages/KScorrect/index.html}{\code{KScorrect}} package.
The approach has been generalised in three ways:
\itemize{
\item General distributions are supported, rather than the closed list used by \code{KScorrect}.
\item Multiple statistical tests are supported, not just \href{https://en.wikipedia.org/wiki/Kolmogorov\%E2\%80\%93Smirnov_test}{KS}.
\item Testing can be performed against distributions fitted to overlapping data, not just IID data,
using the idea of a Gaussian copula to induce autocorrelation consistent with overlapping data suggested in section 4.2 of the
\href{https://www.cambridge.org/core/journals/british-actuarial-journal/article/calibration-of-var-models-with-overlapping-data/B20D66D81DB918AFD3BBDF9EDAC20863}{2019 paper}
by the \href{https://www.actuaries.org.uk/practice-areas/life/research-working-parties/extreme-events}{Extreme Events Working Party}
of the UK \href{https://www.actuaries.org.uk/}{Institute and Faculty of Actuaries}.
}

The function \code{\link{gof_test_sim}} is the main entry point and abstracts some of the implementation details.
\code{\link{gof_test_sim_uniparam}} is the lower-level function that performs the calculations.
}
